Remade the app UI entirely, added a back button to MonthlyTable to go back to the Homepage, and remade the app toolbar | Fixed a major bug that happened when the user selected a month and year that was not previously created and clicked continue from the MonthPicker

+++ backend/laravel/app/Http/Controllers/FinanceController.php
@@ -11,6 +11,12 @@ class FinanceController extends Controller
 {
      public function delete(Request $request, $itemId) {
         try { 
+
+            //Item id is always 0 when the item is only created in frontend and was never saved to database
+            if($itemId == 0) {
+                return Response::json(["Finance wasnt saved to database, therefore no need to deletion, but no problems found"], 200);
+            }
+
             $finance = Finance::firstOrFail();
             $finance->delete();
         } catch (Exception $e) {
